requirements.txt
amqp==5.1.1
astroid==2.11.7
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dill==0.3.5.1
Flask==2.2.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pylint==2.14.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
tomlkit==0.11.1
tweepy==4.10.0
typing-extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
